requirements.txt
python-telegram-bot==20.7
gspread==5.12.4
google-auth==2.27.0
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.5
python-dotenv==1.0.1
pydub==0.25.1

